+++ src/queries.rs
@@ -15,3 +15,5 @@ pub const ZIG_TAGS_QUERY: &str = include_str!("../queries/zig.scm");
 pub const SWIFT_TAGS_QUERY: &str = include_str!("../queries/swift.scm");
 
 pub const TERRAFORM_TAGS_QUERY: &str = include_str!("../queries/terraform.scm");
+
+pub const OBJECTIVE_C_TAGS_QUERY: &str = include_str!("../queries/objective_c.scm");