+++ CHANGELOG.md
@@ -20,6 +20,7 @@ Allow users to manage the downloading of WASM grammars not bundled with treetags
 - Add basic tags support for Zig
 - Add basic tags support for Terraform
 - Add basic tags support for Swift
+- Add basic tags support for Objective C
 
 ### Changed
 - Tag Address is truncated to 96 chars for native parsers